requirements.txt
sqlalchemy==1.4.47
snowflake-sqlalchemy
langchain==0.0.166
#langchain==0.0.190
sqlalchemy-aurora-data-api
PyAthena[SQLAlchemy]==2.25.2
anthropic
redshift-connector==2.0.910
sqlalchemy-redshift==0.8.14

